+++ src/lib/components/pollbuilder/MainPollSettings.svelte
