+++ src/lib/components/pollbuilder/PollSectionSettings.svelte
